Use block_padding for PKCS#7 unpadding of YK HMAC challenge data
This PKCS#7 padding version is possibly incompatible with Yubikey's PKCS#7 version, as it expects
the last byte to always be the padding byte value. See KeepassXC implementation comments
for the details.
Everything works with the challenge length up to 63 bytes though, and YK implementation
would not handle more anyway, hence accepting this potential incompatibility.